Revert "Feat/discovery"
