+++ apps/backend/src/app.module.ts
@@ -5,9 +5,8 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { EventEmitterModule } from '@nestjs/event-emitter';
 import { TaskerooModule } from './taskeroo/taskeroo.module';
 import { WikirooModule } from './wikiroo/wikiroo.module';
-import { McpRegistryModule } from './mcp-registry/mcp-registry.module';
 import { AuthorizationServerModule } from './authorization-server/authorization-server.module';
-import { DiscoveryModule } from './discovery/discovery.module';
+import { McpRegistryModule } from './mcp-registry/mcp-registry.module';
 
 @Module({
   imports: [
@@ -20,9 +19,8 @@ import { DiscoveryModule } from './discovery/discovery.module';
     EventEmitterModule.forRoot(),
     TaskerooModule,
     WikirooModule,
-    McpRegistryModule,
     AuthorizationServerModule,
-    DiscoveryModule,
+    McpRegistryModule,
   ],
   controllers: [AppController],
   providers: [AppService],
