+++ apps/backend/src/authorization-server/authorization-server.module.ts
@@ -3,12 +3,13 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { ClientRegistrationService } from './client-registration.service';
 import { ClientRegistrationController } from './client-registration.controller';
 import { RegisteredClientEntity } from './registered-client.entity';
-
+import { AuthorizationServerMetadataController } from './authorization-server-metadata.controller';
+import { McpRegistryModule } from '../mcp-registry/mcp-registry.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([RegisteredClientEntity])],
+  imports: [TypeOrmModule.forFeature([RegisteredClientEntity]), McpRegistryModule],
   providers: [ClientRegistrationService],
-  controllers: [ClientRegistrationController],
+  controllers: [ClientRegistrationController, AuthorizationServerMetadataController],
   exports: [ClientRegistrationService],
 })
 export class AuthorizationServerModule {}
