--- apps/backend/src/discovery/discovery.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { DiscoveryController } from './discovery.controller';
-import { McpRegistryModule } from '../mcp-registry/mcp-registry.module';
-
-@Module({
-  imports: [McpRegistryModule],
-  controllers: [DiscoveryController],
-})
-export class DiscoveryModule {}
