+++ apps/ui/src/mcp-registry/api.ts
@@ -1,11 +1,6 @@
 import { OpenAPI, McpRegistryService } from 'shared';
 
 // Use relative URL in production, absolute URL in development
-if (import.meta.env.PROD) {
-  OpenAPI.BASE = '';
-} else {
-  const PORT = import.meta.env.VITE_BACKEND_PORT || 3000;
-  OpenAPI.BASE = `http://localhost:${PORT}`;
-}
+OpenAPI.BASE = import.meta.env.PROD ? '' : 'http://localhost:3000';
 
 export { McpRegistryService };
